+++ internal/core/flag.go
@@ -25,10 +25,11 @@ func DeprecatedFlagOption() FlagOptionFunc {
 	return func(cmd *Command, flagName string) {
 		cmd.Command.Flag(flagName).Usage = fmt.Sprintf("%s (deprecated)", cmd.Command.Flag(flagName).Usage)
 		// For documentation purposes, add flag to command Annotation
-		cmd.Command.Annotations = map[string]string{DeprecatedFlagsAnnotation: fmt.Sprintf(flagNamePrintF,
-			cmd.Command.Annotations[DeprecatedFlagsAnnotation],
-			flagName,
-			strings.ToUpper(strings.ReplaceAll(flagName, "-", "_")))}
+		if len(cmd.Command.Annotations) > 0 {
+			cmd.Command.Annotations[DeprecatedFlagsAnnotation] = fmt.Sprintf(flagNamePrintF, cmd.Command.Annotations[DeprecatedFlagsAnnotation], flagName, strings.ToUpper(strings.ReplaceAll(flagName, "-", "_")))
+		} else {
+			cmd.Command.Annotations = map[string]string{DeprecatedFlagsAnnotation: fmt.Sprintf(flagNamePrintF, cmd.Command.Annotations[DeprecatedFlagsAnnotation], flagName, strings.ToUpper(strings.ReplaceAll(flagName, "-", "_")))}
+		}
 	}
 }
 
@@ -36,10 +37,11 @@ func RequiredFlagOption() FlagOptionFunc {
 	return func(cmd *Command, flagName string) {
 		cmd.Command.Flag(flagName).Usage = fmt.Sprintf("%s (required)", cmd.Command.Flag(flagName).Usage)
 		// For documentation purposes, add flag to command Annotation
-		cmd.Command.Annotations = map[string]string{RequiredFlagsAnnotation: fmt.Sprintf(flagNamePrintF,
-			cmd.Command.Annotations[RequiredFlagsAnnotation],
-			flagName,
-			strings.ToUpper(strings.ReplaceAll(flagName, "-", "_")))}
+		if len(cmd.Command.Annotations) > 0 {
+			cmd.Command.Annotations[RequiredFlagsAnnotation] = fmt.Sprintf(flagNamePrintF, cmd.Command.Annotations[RequiredFlagsAnnotation], flagName, strings.ToUpper(strings.ReplaceAll(flagName, "-", "_")))
+		} else {
+			cmd.Command.Annotations = map[string]string{RequiredFlagsAnnotation: fmt.Sprintf(flagNamePrintF, cmd.Command.Annotations[RequiredFlagsAnnotation], flagName, strings.ToUpper(strings.ReplaceAll(flagName, "-", "_")))}
+		}
 	}
 }